🚚  Move Token and TokenType to core
