+++ packages/ntm_ast/lib/ntm_ast.dart
@@ -3,5 +3,3 @@ library ntm_ast;
 export 'src/ast_printer.dart';
 export 'src/expression.dart';
 export 'src/statement.dart';
-export 'src/token.dart';
-export 'src/token_type.dart';
