+++ packages/ntm_ast/test/ast_printer_test.dart
@@ -1,4 +1,5 @@
 import 'package:ntm_ast/ntm_ast.dart';
+import 'package:ntm_core/ntm_core.dart';
 import 'package:test/test.dart';
 
 void main() {
