+++ packages/ntm_core/lib/ntm_core.dart
@@ -2,3 +2,5 @@ library ntm_core;
 
 export 'src/describable.dart';
 export 'src/describable_error.dart';
+export 'src/token.dart';
+export 'src/token_type.dart';
