+++ packages/ntm_interpreter/lib/src/environment.dart
@@ -1,4 +1,4 @@
-import 'package:ntm_ast/ntm_ast.dart';
+import 'package:ntm_core/ntm_core.dart';
 
 import 'runtime_error.dart';
