+++ packages/ntm_scanner/lib/src/scanner.dart
@@ -1,4 +1,4 @@
-import 'package:ntm_ast/ntm_ast.dart';
+import 'package:ntm_core/ntm_core.dart';
 
 import 'scanner_error.dart';
